pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.14.1",
    "duckduckgo-search>=8.1.1",
    "ipython>=9.10.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.9",
    "python-dotenv>=1.2.2",
    "wikipedia>=1.4.0",
]
